+++ test/command_runner_test.dart
@@ -54,8 +54,26 @@ Available commands:
 Run "test help <command>" for more information about a command."""));
     });
 
-    test("supports newlines in command descriptions", () {
+    test("truncates newlines in command descriptions by default", () {
       runner.addCommand(new MultilineCommand());
+
+      expect(runner.usage, equals("""
+A test command runner.
+
+Usage: test <command> [arguments]
+
+Global options:
+-h, --help    Print this usage information.
+
+Available commands:
+  help        Display help information for test.
+  multiline   Multi
+
+Run "test help <command>" for more information about a command."""));
+    });
+
+    test("supports newlines in command summaries", () {
+      runner.addCommand(new MultilineSummaryCommand());
 
       expect(runner.usage, equals("""
 A test command runner.